дз3

+++ seminar01/cocoa.java
@@ -1,7 +1,7 @@
 package seminar01;
 
 public class cocoa extends hotdrink {
-    public cocoa(String name, double price, int volume, int temp) {
+    public cocoa(String name, long price, int volume, int temp) {
         super(name, price, volume, temp);
     }
 }
