+++ seminar01/coffee.java
@@ -1,7 +1,7 @@
 package seminar01;
 
 public class coffee extends hotdrink {
-    public coffee(String name, double price, int volume, int temp) {
+    public coffee(String name, long price, int volume, int temp) {
         super(name, price, volume, temp);
     }
 }
