+++ seminar01/tea.java
@@ -1,7 +1,7 @@
 package seminar01;
 
 public class tea extends hotdrink {
-    public tea(String name, double price, int volume, int temp) {
+    public tea(String name, long price, int volume, int temp) {
         super(name, price, volume, temp);
     }
 }
